+++ src/LoadingIndicator.WinForms/LoadingIndicator.Winforms.csproj
@@ -6,16 +6,16 @@
     <Copyright>Copyright ©  2021</Copyright>
     <PackageVersion>1.1.0-beta</PackageVersion>
     <AssemblyVersion>1.1.0.0</AssemblyVersion>
-    <FileVersion>1.1.0.0</FileVersion>    
+    <FileVersion>1.1.0.0</FileVersion>
     <Authors>Maxim Shoshin</Authors>
     <PackageIcon>icon.png</PackageIcon>
-    <PackageId>LoadingIndicator.WinForms</PackageId>  
-    <PackageProjectUrl>https://github.com/MaxShoshin/LoadingIndicator</PackageProjectUrl>    
+    <PackageId>LoadingIndicator.WinForms</PackageId>
+    <PackageProjectUrl>https://github.com/MaxShoshin/LoadingIndicator</PackageProjectUrl>
     <RepositoryUrl>https://github.com/MaxShoshin/LoadingIndicator</RepositoryUrl>
     <PackageTags>Winforms, Loading, async, indicator, long, operation, windows.forms</PackageTags>
     <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
     <Title>LoadingIndicator.WinForms</Title>
-    <TargetFrameworks>net45;net5.0-windows</TargetFrameworks>
+    <TargetFrameworks>net45;net5.0-windows;net6.0-windows</TargetFrameworks>
     <LangVersion>latest</LangVersion>
     <UseWindowsForms>true</UseWindowsForms>
     <IsPackable>true</IsPackable>
@@ -29,11 +29,11 @@
 
   <ItemGroup>
     <Content Include="LICENSE.txt" Pack="true" PackagePath="LICENSE.txt" />
-    <None Include="icon.png" Pack="true" PackagePath="\"/>
+    <None Include="icon.png" Pack="true" PackagePath="\" />
   </ItemGroup>
-  
-  <ItemGroup> 
-    <PackageReference Include="JetBrains.Annotations" Version="2018.3.0" >
+
+  <ItemGroup>
+    <PackageReference Include="JetBrains.Annotations" Version="2021.3.0">
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
   </ItemGroup>
